+++ go.work
@@ -3,4 +3,5 @@ go 1.21.4
 use (
 	.
 	./tools/format_self
+	./tools/format_unittests
 )
