+++ tools/format_unittests/go.mod
@@ -1,3 +1,7 @@
 module github.com/git-town/git-town/tools/format_unittests
 
 go 1.21.4
+
+require github.com/shoenig/test v1.7.0
+
+require github.com/google/go-cmp v0.6.0 // indirect
